[Imports] Remove unnecessary local imports
Break the constraint <-> collision_driver <-> solver circular dependency
by moving constraint_sizes() into collision_driver.py (where its only
non-type dependency, make_condim, already lives). This lets constraint.py
import collision_driver at module level instead of inside every function.

Also promote remaining local imports to module level:
- device.py: tree_map from torch.utils._pytree
- support.py: math module, functorch internals
- solver.py: constraint_sizes (now from collision_driver)
